requirement.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
brotli==1.2.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.7
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.2
django-sendgrid-v5==1.3.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
gunicorn==23.0.0
h11==0.16.0
idna==3.11
inflection==0.5.1
kombu==5.5.4
MarkupSafe==3.0.3
mysqlclient==2.2.7
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.3
rabbitmq==0.2.0
redis==6.4.0
requests==2.32.5
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
whitenoise==6.11.0
